Skip creneau checking when motif.follow_up? and user not logged in

+++ app/controllers/concerns/can_have_rdv_wizard_context.rb
@@ -13,6 +13,7 @@ module CanHaveRdvWizardContext
 
     parsed_params = Rack::Utils.parse_nested_query(parsed_uri.query).to_h.symbolize_keys
     rdv_wizard = UserRdvWizard::Step1.new(nil, parsed_params)
+    return if rdv_wizard.motif.follow_up?
 
     if rdv_wizard.creneau.blank?
       session.delete(:user_return_to)
